+++ docs/changelog.md
@@ -5,6 +5,8 @@
 ### Fixed
 
 - Fixed a bug in `pint ci` that would cause a failure if a directory was renamed.
+- Fixed false positive reports from `promql/series` check when a time series
+  disappears from Prometheus.
 
 ## v0.40.0